// web-angular/src/app/model/individual.ts
export class Individual{
	individualId: number;
	createBy: String;
	createDate: Date;
	dateOfBirth: Date;
	debiturName: String;
	identityNumber: number;
	modifyBy: String;
	modifyDate: Date;
	npwp: String;
	placeOfBirth: String;
	refUserCode: String;
	reqPurposeId: number;
	sexId: number;
	status: number;
	statusSend: number;
	combinationType: number;
	searchType: number;
	sourceData: number;
	typeInteractive: String;
	startDate: Date;
	finishedDate: Date;
}